rag_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sentence-transformers==3.0.1
qdrant-client==1.10.1
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv==1.0.1
numpy>=1.26.0
httpx==0.27.0